[profiler] Change sample frequency to a 32-bit int.
The unit is hertz, so values above 100-1000 are already extremely unusual.
Changing it to a 32-bit int avoids word tearing on 32-bit platforms.


Commit migrated from https://github.com/mono/mono/commit/6214c2e6edf147dad216dad8fa45c11dca1b4a17

+++ src/mono/mono/profiler/log.c
@@ -4184,7 +4184,7 @@ proflog_icall_SetCallDepth (gint32 value)
 ICALL_EXPORT void
 proflog_icall_GetSampleMode (MonoProfilerSampleMode *mode, gint32 *frequency)
 {
-	guint64 freq;
+	uint32_t freq;
 
 	mono_profiler_get_sample_mode (log_profiler.handle, mode, &freq);
